requirements.txt
attrs==23.1.0
barcode==1.0.3
brother-ql==0.9.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.45.1
future==0.18.3
matplotlib==3.8.2
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
packbits==0.6
pandas==2.2.0
pdf417==0.8.1
Pillow==10.1.0
pyparsing==3.1.1
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.8.2
python-Levenshtein==0.23.0
pytz==2024.1
pyusb==1.2.1
qrcode==7.4.2
rapidfuzz==3.5.2
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
